Update README.md
# ccjmne.github.io

My personal website.  
Not intended as a portfolio nor an online curriculum vitæ, but as a mere playground (albeit organised) for me to quench my thirst for experimentation with the ever-new Vanilla ECMAScript and other platform-native Web front-end goodies.  
`#UseThePlatform`

Developed for Chromium.

Built with a bit of ❤ and a lot of Functional Programming.

## Contributing

Sure! 🤔

## Licensing

This project embraces the GNU General Public License v3.0.

You may copy, distribute and modify the software as long as you track changes/dates in source files. Any modifications to or software including (via compiler) GPL-licensed code must also be made available under the GPL along with build & install instructions.

You can find the precise terms and conditions for copying, distribution and modification in the [LICENSE](LICENSE) file.